LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/history-state.window-expected.txt


PASS history.state is kept by document.open()
PASS history.state is kept by document.open() (with historical replace parameter set)
